修改mapper中的返回值类型

+++ autoJava-demo/src/main/java/com/ja/autojavademo/mappers/ProductInfoMapper.java
@@ -16,12 +16,12 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Id更新
 	 */
-	Integer updateById(@Param("bean") T t, @Param("id") Integer id);
+	Long updateById(@Param("bean") T t, @Param("id") Integer id);
 
 	/**
 	 * 根据Id删除
 	 */
-	Integer deleteById(@Param("id") Integer id);
+	Long deleteById(@Param("id") Integer id);
 
 	/**
 	 * 根据Code查询
@@ -31,12 +31,12 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Code更新
 	 */
-	Integer updateByCode(@Param("bean") T t, @Param("code") String code);
+	Long updateByCode(@Param("bean") T t, @Param("code") String code);
 
 	/**
 	 * 根据Code删除
 	 */
-	Integer deleteByCode(@Param("code") String code);
+	Long deleteByCode(@Param("code") String code);
 
 	/**
 	 * 根据SkuType和ColorType查询
@@ -46,11 +46,11 @@ public interface ProductInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据SkuType和ColorType更新
 	 */
-	Integer updateBySkuTypeAndColorType(@Param("bean") T t, @Param("skuType") Integer skuType, @Param("colorType") Integer colorType);
+	Long updateBySkuTypeAndColorType(@Param("bean") T t, @Param("skuType") Integer skuType, @Param("colorType") Integer colorType);
 
 	/**
 	 * 根据SkuType和ColorType删除
 	 */
-	Integer deleteBySkuTypeAndColorType(@Param("skuType") Integer skuType, @Param("colorType") Integer colorType);
+	Long deleteBySkuTypeAndColorType(@Param("skuType") Integer skuType, @Param("colorType") Integer colorType);
 
 }
